Extract core work graph CLI commands
